Starting to refactor pixel classifier UI
